Converted project to workspace members
Prerequisite to allow reuse of logic for later GUI fronted
