+++ .github/workflows/build.yml
@@ -40,5 +40,5 @@ jobs:
         run: cargo clippy -- -Dwarnings
         
       - name: "Checking if all unit tests pass" 
-        run: cargo test --verbose
+        run: cargo test --workspace --verbose
