+++ Cargo.toml
@@ -1,59 +1,8 @@
-[package]
-name = "usermgmt"
-version = "0.5.1"
-edition = "2021"
-license-file = "LICENSE"
-description = "Simultaneous Slurm and LDAP user management"
-readme = "README.md"
-rust-version = "1.70.0"
+[workspace]
+members = ["usermgmt", "usermgmt"]
+default-members = ["usermgmt"]
+resolver = "2"
 
-[dependencies]
-toml = "0.8.1"
-clap = { version = "4.4.6", features = ["derive"] }
-serde = { version = "1.0.188", features = ["derive"] }
-confy = "0.5.1"
-tempfile = "3.8.0"
-rpassword = "7.2.0"
+[workspace.dependencies]
 log = "0.4.16"
-env_logger = "0.10.0"
-ldap3 = "0.11.3"
-maplit = "1.0.2"
-ssh2 = "0.9.4"
-once_cell = "1.18.0"
-anyhow = { version = "1.0.75", features = ["backtrace"] }
-prettytable = "0.10.0"
-getset = "0.1.2"
-dirs = "5.0.1"
-derive_more = { version = "0.99.17", default-features = false, features = [
-    "from",
-    "into",
-    "as_ref",
-    "display",
-    "into_iterator",
-] }
-
-[dev-dependencies]
-insta = { version = "1.33.0", features = ["yaml"] }
-
-[package.metadata.deb]
-maintainer = "Dominik Wagner <dominik.wagner@th-nuernberg.de>"
-copyright = "2023, Technische Hochschule Nürnberg Georg Simon Ohm"
-license-file = ["LICENSE", "4"]
-extended-description = """\
-A simple commandline tool for \
-managing Slurm and LDAP users simultaneously."""
-depends = "$auto"
-section = "utility"
-priority = "optional"
-assets = [
-    [
-        "target/release/usermgmt",
-        "usr/bin/",
-        "755",
-    ],
-    [
-        "README.md",
-        "usr/share/doc/usermgmt/README",
-        "644",
-    ],
-]
+clap = { version = "4.4.6", features = ["derive"] }
