+++ usermgmt_lib/src/cli.rs
