+++ usermgmt_lib/src/entity.rs
