+++ usermgmt_lib/src/slurm.rs
