+++ usermgmt_lib/src/slurm/commmand_builder.rs
