+++ usermgmt_lib/src/util.rs
