+++ usermgmt_lib/src/util/result_accumalator.rs
