+++ usermgmt_lib/src/util/trimmed_non_empty_text.rs
